refactor: move LoginMember's package member to auth
