+++ src/main/java/nextstep/member/application/UserDetailsServiceImpl.java
@@ -2,7 +2,7 @@ package nextstep.member.application;
 
 import nextstep.auth.service.UserDetails;
 import nextstep.auth.service.UserDetailsService;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.service.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
@@ -19,6 +19,6 @@ public class UserDetailsServiceImpl implements UserDetailsService {
     @Override
     public UserDetails loadUserByUsername(String email) {
         Member member = memberRepository.findByEmail(email).orElseThrow(RuntimeException::new);
-        return LoginMember.of(member);
+        return LoginMember.of(member.getEmail(), member.getPassword(), member.getRoles());
     }
 }
